+++ Pod/Classes/ViewControllers/ARArtistComponentViewController.m
@@ -9,9 +9,10 @@
 
 - (instancetype)initWithArtistID:(NSString *)artistID emission:(AREmission *)emission;
 {
+  BOOL isPad = UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPad;
   if ((self = [super initWithEmission:emission
                            moduleName:@"Artist"
-                    initialProperties:@{ @"artistID": artistID }])) {
+                    initialProperties:@{ @"artistID": artistID, @"isPad": @(isPad) }])) {
     _artistID = artistID;
   }
   return self;