+++ src/lib/Components/Inbox/Conversations/index.tsx
@@ -84,7 +84,7 @@ export class Conversations extends React.Component<Props, State> {
     }
     this.setState({ fetchingNextPage: true })
     this.props.relay.loadMore(PageSize, error => {
-      // TODO Not performing any error handling here
+      // TODO: Not performing any error handling here
       this.setState({
         fetchingNextPage: false,
         dataSource: this.state.dataSource.cloneWithRows(this.conversations),
@@ -163,9 +163,9 @@ export default createPaginationContainer(
       }
     },
     query: graphql.experimental`
-      query ConversationsQuery($count: Int!, $after: String) {
+      query ConversationsQuery($count: Int!, $cursor: String) {
         me {
-          ...Conversations_me @arguments(count: $count, after: $after)
+          ...Conversations_me @arguments(count: $count, cursor: $cursor)
         }
       }
     `,